Split Android database migration tests
